requirements.txt
# Core
pydantic>=2.0.0
loguru>=0.7.0

# Signal processing
scipy>=1.11.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Database
# sqlite3 is stdlib — no extra package needed

# Development
black>=23.0.0
ruff>=0.0.290
mypy>=1.5.0

# Webcam (Step 2)
mediapipe>=0.10.0
opencv-python>=4.8.0

# Knowledge graph (Step 3)
neo4j>=5.0.0

# NLP Pipeline (Step 4)
textblob>=0.18.0

# GPT-4 Intervention Engine (Step 6)
openai>=1.0.0

# Content Generation Pipeline (Step 7)
pdfplumber>=0.9.0
python-pptx>=0.6.21
moviepy>=1.0.3
pillow>=10.0.0

# LangGraph Fusion Engine (Step 5)
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# Spaced Repetition Engine (Step 8)
plyer>=2.1.0

# Desktop Application API (Step 11)
fastapi>=0.104.0

# Experiment Framework (Step 12)
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
reportlab>=4.0.0
jinja2>=3.1.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Future steps (commented out — install when needed)
# pyserial>=3.5            # Step 10
# mne>=1.5.0               # Step 10
